Update README.md
Basic personal portfolio website. It has five sections: Home,About,Project,Services and Contact. Each section has its own provision and basic info about it. It is a responsive website, that is, it works for mobile sizes as well. When we click on a specific tab, it scrolls us to that version. There is a downloadable file of resume/CV provided as well. There is a javascript validation provided for the "submit" form present in the contact form. It stores all the data of the viewer/person who sent a message to my email in the Excel Sheets. It has html,css and javascript code.